requirements.txt
swibots
beautifulsoup4
aiohttp
playwright
user_agent
python-decouple